Update README.md
Mastocytosis is a rare skin disease that is caused by mast cells that generate a chemical called histamine. Mast cells are usually found in the bone marrow of the human body and they are part of our immune system. Most of the mast cells are concentrated on our skins, lungs, and intestine and are also found in many other parts of our body. Mast cells are responsible for maintaining physiological functions as well as prevent the entry of pathogens.

Most of the mastocytosis mutations occur due to the KIT gene on the 4q12 chromosome which increases cell production and causes an overgrowth of one cell line in the mast cells. When the mast cell is activated due to the excessive release of these chemicals it causes the blood vessels to leak and results in causing localized itching, swelling, redness, and blisters. Currently, the mastocytosis that affects human skin is classified into Cutaneous and Systemic mastocytosis. And since the characteristics of mast cell activation syndrome look similar to the characteristics of other common skin diseases. Doctors are required to perform numerous tests on the affected site to detect the presence of mast cells. This can be a challenging task in developing countries since most of the doctors have limited access to test facilities and funding from the government.

Once the type of mastocytosis on the skin is identified easily it might help the doctors in skipping the tedious steps in discovery and help them concentrate more on treating the syndrome. We as data science graduates wanted to find an effective way that might help the doctors in identifying the type of mastocytosis with the help of two-photon fluorescence lifetime imaging.

Currently, the advancements in image analysis with the help of ML algorithms have grown exponentially and numerous algorithms can be integrated to train a model step by step in achieving accuracy in detection. We wanted to find out how can an ML model predict the type of mastocytosis with the use of existing algorithms.